move authenticate to middleware
